improve

+++ fetch-debian-repo.sh
@@ -1,12 +1,23 @@
 #!/bin/bash
 
+# arguments
+usage() {
+	echo "$0 <distro> <release>"
+}
+if [ $# != 2 ]; then
+	usage
+	exit 1
+fi
+DISTRO=$1
+RELEASE=$2
+
 # settings
 MIRROR=http://ftp.no.debian.org/
-DISTRO=debian
-RELEASE=wheezy
+#DISTRO=debian
+#RELEASE=wheezy
 REPO=main
 ARCHITECTURE=armhf
-DESTDIR=dl
+DESTDIR="$PWD"
 
 ##############################################################################################
 
@@ -51,19 +62,30 @@ function download_checked() {
 	while IFS= read line; do
 		chksum=`echo $line | cut -d' ' -f2`
 		url=`echo $line | cut -d' ' -f1`
-
 		filename=`basename $url`
+		printf "%s: " "$filename"
+
+		changed=no
 		if [ -e "$filename" ]; then
-			echo "$chksum $filename" | md5sum -c
+			echo "$chksum $filename" | md5sum -c 2>/dev/null 1>/dev/null
 			if [ $? -eq 0 ]; then
+				printf "skipped\n"
 				continue
 			fi
-			rm -fv $filename
+			rm -f $filename
+			changed=yes
 		fi
 
 		curl -k $url -o $filename 2>/dev/null
 		if [ $? -ne 0 ]; then
-			echo "Failed to download $url!"
+			printf "failed\n"
+			#echo "Failed to download $url!"
+		else
+			if [ "x$changed" = "xno" ]; then
+				printf "done\n"
+			else
+				printf "updated\n"
+			fi
 		fi
 	done
 }
@@ -73,12 +95,16 @@ function download_checked() {
 mkdir -p $DESTDIR
 pushd $DESTDIR
 
-# download release file for the record
-curl -k $MIRROR/$DISTRO/dists/$RELEASE/$REPO/binary-$ARCHITECTURE/Release -o Release
+if [ ! -e Release ] || [ ! -e Packages ]; then
+	rm -f Release Packages
 
-# download packages list
-curl -k $MIRROR/$DISTRO/dists/$RELEASE/$REPO/binary-$ARCHITECTURE/Packages.bz2 -o Packages.bz2
-bunzip2 Packages.bz2
+	# download release file for the record	
+	curl -k $MIRROR/$DISTRO/dists/$RELEASE/$REPO/binary-$ARCHITECTURE/Release -o Release
+
+	# download packages list
+	curl -k $MIRROR/$DISTRO/dists/$RELEASE/$REPO/binary-$ARCHITECTURE/Packages.bz2 -o Packages.bz2
+	bunzip2 Packages.bz2
+fi
 
 # parse packages file and download packages
 parse_pkgs Packages | download_checked
